Add Node struct to Tree to manage child Problems.
Each Node in Tree has an int heuristics_cost and int actions_costs
attributes, which are set when the state is expanded into a new Problem.
These ints propagate the next states to expand based on the search
algorithm in use.

Removed Node class and header to clean up the codebase.