Change move_zero_tile() to take in char instead of 2 ints.
In Tree, the search functions check for legal moves using
move_zero_tile(). Checking for legal moves requires moving up, left,
down, or right in a 2D array. move_zero_tile() takes in (-1, 0), (0, -1),
(1, 0), or (0, 1) respectively, which is difficult to iterate through.

Instead, define the possible moves in move_zero_tile(), and the function
determines which move to use based on its parameter of 'u', 'l', 'd',
or 'r'. This makes clear what direction the zero tile is moving in
the multiple usages of move_zero_tile() in Tree.