Implement function to perform uniform cost search.
Added Node vector frontier_, array of 4 Node pointers, hashmap explored_map_,
a const char array of moves_[4] as private members in Tree to use in
multiple search functions.

Function uniform_cost_search() uses a vector to keep track of nodes
to expand. When a node is expanded and possible problem states are found,
possible states are converted to a string and looked for in the hashmap.
If the state does not exist in the hashmap, the state is not explored,
and the state is added to a queue. Once all possible problem states are
explored, nodes in the queue are sorted by least to greatest cost into
the vector.